Corrections
Corrections for Maier, changed prototype. In Maier,
the special case where we delete an implication at index i and modify imp at j < i was a problem. 
To solve it, we introduced an array T containing for each implication, the number of times it has been moved by one to the left. thus, j in the min basis will be at the position (j - T[j). This is a pb only when j < i, because we remove implication in order.